WIP: Add asymetricThinningSchemeWithSplits
Change ConstAlias to Clone in CubicalComplex constructor with kspace
and raw pointer KSpace* to CowPtr<KSpace> to manage memory
automatically.

WIP: There was a domain error when a voxel was in the border
with K_0,1,2,3 functions. Now it uses PreCells.
Add belongs(PreCell &) to CubicalComplex that returns false
if the PreCell is outside of the kspace domain.
TODO: This might have introduced a bug