Add a domainShift in ImageContainerByITKImage
Allowing to shift the domain, without affecting the indexing
of the underlying ITK image.

Also add basic conversions from index points to physical points.

The `domainShift` parameter can only be set with the existing
function`updateDomain`. It defaults to `Zero`.

The implementation provided in #1484 was untested,
and I would say it goes out of the buffer of the
ITK region when using `operator()` or `setValue`.

This patch introduces a conversion function between what I call a
`domainPoint` (a point in the DGtal domain defined by `lowerBound` and `upperBound`),
and an `indexPoint` which is a valid point in the ITK region.

These two are equivalent when `domainShift` is Zero (the default).

It also effectively reverts #1484 that mixes index space and physical
space (with origins) in ITK.